(BSR)[PRO] chore: add unit test on table
