--- pro/src/pages/IndividualOffers/IndividualOffersContainer/components/IndividualOffersTable/components/IndividualOffersTableBody/IndividualOffersTableBody.module.scss
@@ -1,11 +1,0 @@
-@use "styles/mixins/_size.scss" as size;
-
-.individual-tbody {
-  display: block;
-}
-
-@media (min-width: size.$tablet) {
-  .individual-tbody {
-    display: table-row-group;
-  }
-}
